Add no key hashing option to ProofMapIndexProxy [ECR-3779] (#1222)
