+++ exonum-java-binding/core/rust/src/storage/proof_map_index.rs
@@ -63,6 +63,8 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofMapInde
     _: JClass,
     name: JString,
     view_handle: Handle,
+    // TODO: to be used in ECR-3765
+    _key_hashing: jboolean,
 ) -> Handle {
     let res = panic::catch_unwind(|| {
         let name = utils::convert_to_string(&env, name)?;
@@ -88,6 +90,8 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofMapInde
     group_name: JString,
     map_id: jbyteArray,
     view_handle: Handle,
+    // TODO: to be used in ECR-3765
+    _key_hashing: jboolean,
 ) -> Handle {
     let res = panic::catch_unwind(|| {
         let group_name = utils::convert_to_string(&env, group_name)?;
