+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/storage/indices/MapIndex.java
@@ -94,22 +94,23 @@ public interface MapIndex<K, V> extends StorageIndex {
   void remove(K key);
   
   /**
-   * Returns an iterator over the map keys in lexicographical order.
+   * Returns an iterator over the map keys. The keys are ordered in lexicographical order.
    *
    * @throws IllegalStateException if this map is not valid
    */
   Iterator<K> keys();
 
   /**
-   * Returns an iterator over the map values in lexicographical order of <em>keys</em>.
+   * Returns an iterator over the map values. The values are ordered in lexicographical order of
+   * <em>keys</em>.
    *
    * @throws IllegalStateException if this map is not valid
    */
   Iterator<V> values();
 
   /**
-   * Returns an iterator over the map entries.
-   * The entries are ordered by keys in lexicographical order.
+   * Returns an iterator over the map entries. The entries are ordered by keys in lexicographical
+   * order.
    *
    * @throws IllegalStateException if this map is not valid
    */
